+++ pythonik/specs/collection.py
@@ -4,12 +4,15 @@ from pythonik.models.assets.collections import (
     Collection,
     CollectionContents,
     CollectionContentInfo,
+    Content,
+    AddContentResponse
 )
 
 BASE = "collections"
 GET_URL = BASE + "/{}/"
 GET_INFO = GET_URL + "content/info"
 GET_CONTENTS = GET_URL + "contents"
+POST_CONTENT = GET_CONTENTS
 
 
 class CollectionSpec(Spec):
@@ -122,3 +125,32 @@ class CollectionSpec(Spec):
             BASE, json=body.model_dump(exclude_defaults=exclude_defaults), **kwargs
         )
         return self.parse_response(response, Collection)
+    
+    def add_content(
+        self, collection_id: str, body: Content, exclude_defaults=True, **kwargs
+    ):
+        """Add an object to a collection.
+
+
+        Args:
+            collection_id (str): The ID of the collection to add content to
+            body (Content): The content object containing object_id and object_type
+            exclude_defaults (bool, optional): Whether to exclude default values from the request. Defaults to True.
+            **kwargs: Additional keyword arguments to pass to the request
+
+        Returns:
+            Response[Collection]: The updated collection object
+
+        Response Codes:
+            201: Content added successfully
+            400: Bad request
+            401: Token is invalid
+            404: Collection not found
+        """
+        response = self._post(
+            POST_CONTENT.format(collection_id),
+            json=body.model_dump(exclude_defaults=exclude_defaults),
+            **kwargs,
+        )
+        print(f"response: {response.json()}")
+        return self.parse_response(response, AddContentResponse)